Move CLI to menu
