+++ node_launcher/gui/network_buttons/advanced/__init__.py
@@ -1,5 +1,4 @@
 from .advanced_widget import AdvancedWidget
-from .cli_layout import CliLayout
 from .configuration_files_layout import ConfigurationFilesLayout
 from .ports_layout import PortsLayout
 from .tls_layout import TlsLayout
